convert flex-base.lua into a themepark theme
This already allows to run Nominatim under themepark, currently
as a topic-less theme.
